+++ task_files/go_scripts/npm_switch_to_new_version/main.go
@@ -33,7 +33,6 @@ func main() {
 	}
 	handleSpecialPackages := utils.ShowMenu(cliInfo,nil)
 
-
 	nvmPath, err := findCommandPath("nvm")
 	askForPath := false
 	if err != nil {